fix(vm-list) fix VM cards font color
change the font color by classes
