+++ src/app/shared/services/utils/utils.service.ts
@@ -75,6 +75,17 @@ export class Utils {
       return null;
     }
   }
+
+  public static isColorDark(color: string) {
+    const hex = color.replace('#', '');
+    const r = parseInt(hex.substring(0, 2), 16);
+    const g = parseInt(hex.substring(2, 4), 16);
+    const b = parseInt(hex.substring(4, 6), 16);
+
+    const darkness = 1 - ( 0.299 * r + 0.587 * g + 0.114 * b) / 255;
+
+    return darkness > 0.5;
+  }
 }
