+++ src/app/vm/vm-list/vm-list-item.component.scss
@@ -20,6 +20,12 @@
       color: inherit;
     }
   }
+  &.dark-background {
+    color: rgba(255, 255, 255, 0.87) !important;
+  }
+  &.light-background {
+    color: rgba(0, 0, 0, 0.87) !important;
+  }
   .state-icon {
     box-shadow: 0 0 1px 1px white;
   }
