+++ src/app/vm/vm-list/vm-list-item.component.ts
@@ -12,6 +12,7 @@ import { MdMenuTrigger } from '@angular/material';
 import { Color } from '../../shared/models';
 import { VmTagService } from '../../shared/services/tags/vm-tag.service';
 import { VirtualMachine, VmState } from '../shared/vm.model';
+import { Utils } from '../../shared/services/utils/utils.service';
 
 const stateTranslations = {
   RUNNING: 'VM_STATE.RUNNING',
@@ -90,6 +91,8 @@ export class VmListItemComponent implements OnInit, OnChanges {
     return {
       'card-selected': this.isSelected(this.item),
       'has-text-color': !!this.color && !!this.color.textColor,
+      'dark-background':  !!this.color && Utils.isColorDark(this.color.value),
+      'light-background': !this.color || !Utils.isColorDark(this.color.value),
       error,
       destroyed
     };
@@ -113,20 +116,5 @@ export class VmListItemComponent implements OnInit, OnChanges {
   private updateColor(): void {
     const savedColor = this.vmTagService.getColorSync(this.item);
     this.color = new Color(savedColor.name, savedColor.value);
-
-    this.color.textColor = this.isBackgroundDark(savedColor.value)
-      ? 'rgba(255, 255, 255, 0.87)'
-      : 'rgba(0, 0, 0, 0.87)';
-  }
-
-  private isBackgroundDark(color: string) {
-    const hex = color.replace('#', '');
-    const r = parseInt(hex.substring(0, 2), 16);
-    const g = parseInt(hex.substring(2, 4), 16);
-    const b = parseInt(hex.substring(4, 6), 16);
-
-    const darkness = 1 - ( 0.299 * r + 0.587 * g + 0.114 * b) / 255;
-
-    return darkness > 0.5 ? true : false;
   }
 }
